+++ src/buffer.h
@@ -11,7 +11,7 @@ extern "C" {
 
 typedef struct {
 	unsigned char *ptr;
-	int asize, size;
+	size_t asize, size;
 } cmark_strbuf;
 
 extern unsigned char cmark_strbuf__initbuf[];
@@ -26,7 +26,7 @@ extern unsigned char cmark_strbuf__oom[];
  * For the cases where GH_BUF_INIT cannot be used to do static
  * initialization.
  */
-void cmark_strbuf_init(cmark_strbuf *buf, int initial_size);
+void cmark_strbuf_init(cmark_strbuf *buf, size_t initial_size);
 
 /**
  * Attempt to grow the buffer to hold at least `target_size` bytes.
@@ -36,7 +36,7 @@ void cmark_strbuf_init(cmark_strbuf *buf, int initial_size);
  * existing buffer content will be preserved, but calling code must handle
  * that buffer was not expanded.
  */
-int cmark_strbuf_try_grow(cmark_strbuf *buf, int target_size, bool mark_oom);
+int cmark_strbuf_try_grow(cmark_strbuf *buf, size_t target_size, bool mark_oom);
 
 /**
  * Grow the buffer to hold at least `target_size` bytes.
@@ -46,7 +46,7 @@ int cmark_strbuf_try_grow(cmark_strbuf *buf, int target_size, bool mark_oom);
  *
  * @return 0 on success or -1 on failure
  */
-int cmark_strbuf_grow(cmark_strbuf *buf, int target_size);
+int cmark_strbuf_grow(cmark_strbuf *buf, size_t target_size);
 
 void cmark_strbuf_free(cmark_strbuf *buf);
 void cmark_strbuf_swap(cmark_strbuf *buf_a, cmark_strbuf *buf_b);
@@ -68,9 +68,9 @@ size_t cmark_strbuf_len(const cmark_strbuf *buf);
 
 int cmark_strbuf_cmp(const cmark_strbuf *a, const cmark_strbuf *b);
 
-void cmark_strbuf_attach(cmark_strbuf *buf, unsigned char *ptr, int asize);
+void cmark_strbuf_attach(cmark_strbuf *buf, unsigned char *ptr, size_t asize);
 unsigned char *cmark_strbuf_detach(cmark_strbuf *buf);
-void cmark_strbuf_copy_cstr(char *data, int datasize, const cmark_strbuf *buf);
+void cmark_strbuf_copy_cstr(char *data, size_t datasize, const cmark_strbuf *buf);
 
 static inline const char *cmark_strbuf_cstr(const cmark_strbuf *buf)
 {
@@ -87,20 +87,20 @@ static inline const char *cmark_strbuf_cstr(const cmark_strbuf *buf)
  * return code of these functions and call them in a series then just call
  * cmark_strbuf_oom at the end.
  */
-int cmark_strbuf_set(cmark_strbuf *buf, const unsigned char *data, int len);
+int cmark_strbuf_set(cmark_strbuf *buf, const unsigned char *data, size_t len);
 int cmark_strbuf_sets(cmark_strbuf *buf, const char *string);
 int cmark_strbuf_putc(cmark_strbuf *buf, int c);
-int cmark_strbuf_put(cmark_strbuf *buf, const unsigned char *data, int len);
+int cmark_strbuf_put(cmark_strbuf *buf, const unsigned char *data, long len);
 int cmark_strbuf_puts(cmark_strbuf *buf, const char *string);
 int cmark_strbuf_printf(cmark_strbuf *buf, const char *format, ...)
 CMARK_ATTRIBUTE((format (printf, 2, 3)));
 int cmark_strbuf_vprintf(cmark_strbuf *buf, const char *format, va_list ap);
 void cmark_strbuf_clear(cmark_strbuf *buf);
 
-int cmark_strbuf_strchr(const cmark_strbuf *buf, int c, int pos);
-int cmark_strbuf_strrchr(const cmark_strbuf *buf, int c, int pos);
-void cmark_strbuf_drop(cmark_strbuf *buf, int n);
-void cmark_strbuf_truncate(cmark_strbuf *buf, int len);
+int cmark_strbuf_strchr(const cmark_strbuf *buf, int c, size_t pos);
+long cmark_strbuf_strrchr(const cmark_strbuf *buf, int c, size_t pos);
+void cmark_strbuf_drop(cmark_strbuf *buf, long n);
+void cmark_strbuf_truncate(cmark_strbuf *buf, size_t len);
 void cmark_strbuf_rtrim(cmark_strbuf *buf);
 void cmark_strbuf_trim(cmark_strbuf *buf);
 void cmark_strbuf_normalize_whitespace(cmark_strbuf *s);
